requirements.txt
chromadb==0.4.24
nltk==3.8.1
google-generativeai==0.5.0
fastapi[all]==0.110.1
